First Prototype
This is the first prototype of this project. 

It is the first one I sent to 3D Print. Feel free to make any changes you deem necessary.

Some things I noticed:

Buttons might be unstable
Structure might not be so reliable when you press the trigger.
There is nothing closing the bottom (where the Arduino goes and the microphone cable passes)
Trigger migh be unstable using some switch types